Software/requirements.txt
pyserial
requests
ttkbootstrap
matplotlib

